requirements.txt
pandas>=2.0.3
openpyxl>=3.1.2
numpy>=1.24.3

# Use the official supported SDK for Gemini / Google GenAI
google-genai>=0.1.0

python-dotenv>=1.0.0
requests>=2.28
fuzzywuzzy[speedup]>=0.18
